cleanup some stuff
